chore(server): move & reorg fiabilisation jobs & files
wip


wip
