+++ server/src/jobs/cli.fiabilisation.js
@@ -7,9 +7,42 @@ import { cleanUaisDuplicates } from "./fiabilisation/duplicates/dossiersApprenan
 import { identifyCfdDuplicates } from "./fiabilisation/duplicates/dossiersApprenants-duplicates-cfd/identify.js";
 import { removeDossierApprenantsDuplicates } from "./fiabilisation/duplicates/dossiersApprenants-duplicates/remove-duplicates.js";
 import { analyseFiabiliteDossierApprenantsRecus } from "./fiabilisation/dossiersApprenants/analyse-fiabilite-dossiers-apprenants-recus.js";
-import { createFiabilisationUaiSiretMapping } from "./fiabilisation/uai-siret/create-fiabilisation-uai-siret-mapping/index.js";
-import { updateDossiersApprenantWithFiabilisationUaiSiret } from "./fiabilisation/uai-siret/update-dossiers-apprenants-with-fiabilisation-uai-siret/index.js";
-import { cleanOrganismesNonFiables } from "./fiabilisation/uai-siret/clean-organismes-non-fiables/index.js";
+
+// /**
+//  * Job de création de la collection de mapping fiabilisation UAI SIRET
+//  */
+// cli
+//   .command("create:mapping-fiabilisation-uai-siret")
+//   .description("Création de la collection de mapping pour fiabilisation des UAI SIRET")
+//   .action(() => {
+//     runScript(async () => {
+//       return createFiabilisationUaiSiretMapping();
+//     }, "create-mapping-fiabilisation-uai-siret");
+//   });
+
+// /**
+//  * Job d'application de la fiabilisation UAI SIRET
+//  */
+// cli
+//   .command("apply:fiabilisation-uai-siret")
+//   .description("Application du mapping de fiabilisation des UAI SIRET")
+//   .action(() => {
+//     runScript(async () => {
+//       return updateDossiersApprenantWithFiabilisationUaiSiret();
+//     }, "apply-fiabilisation-uai-siret");
+//   });
+
+// /**
+//  * Job de suppression des organismes non fiables et déplacement des contributeurs & effectifs liés
+//  */
+// cli
+//   .command("clean:organismes-non-fiables")
+//   .description("Suppression des organismes non fiables et déplacement des contributeurs & effectifs liés")
+//   .action(() => {
+//     runScript(async () => {
+//       return cleanOrganismesNonFiables();
+//     }, "fiabilisation-clean-organismes-non-fiables");
+//   });
 
 /**
  * Job d'identification des doublons dans les fichiers CSV de réseaux
@@ -85,42 +118,6 @@ cli
     runScript(async () => {
       return analyseFiabiliteDossierApprenantsRecus();
     }, "analyse-dossiersApprenants-recus");
-  });
-
-/**
- * Job de création de la collection de mapping fiabilisation UAI SIRET
- */
-cli
-  .command("create:mapping-fiabilisation-uai-siret")
-  .description("Création de la collection de mapping pour fiabilisation des UAI SIRET")
-  .action(() => {
-    runScript(async () => {
-      return createFiabilisationUaiSiretMapping();
-    }, "create-mapping-fiabilisation-uai-siret");
-  });
-
-/**
- * Job d'application de la fiabilisation UAI SIRET
- */
-cli
-  .command("apply:fiabilisation-uai-siret")
-  .description("Application du mapping de fiabilisation des UAI SIRET")
-  .action(() => {
-    runScript(async () => {
-      return updateDossiersApprenantWithFiabilisationUaiSiret();
-    }, "apply-fiabilisation-uai-siret");
-  });
-
-/**
- * Job de suppression des organismes non fiables et déplacement des contributeurs & effectifs liés
- */
-cli
-  .command("clean:organismes-non-fiables")
-  .description("Suppression des organismes non fiables et déplacement des contributeurs & effectifs liés")
-  .action(() => {
-    runScript(async () => {
-      return cleanOrganismesNonFiables();
-    }, "fiabilisation-clean-organismes-non-fiables");
   });
 
 cli.parse(process.argv);
